Python3 doesn't have iteritems, use items.
In fact python3 items returns dict_items object opposed to list of tuples
in python2. The dict_items object is iterable and optimized.
Unfortunately, this has again perfomance (memory) impact on python2, but
since this code is in handling traceback part, it shouldn't be run more
then once, so this impact is minimal.